Switch to BigInteger and BigDecimal for journal/eddn numbers since values are unsigned and went out of range add several missing fields

+++ application/src/main/java/nl/jixxed/eliteodysseymaterials/parser/messageprocessor/BackpackChangeMessageProcessor.java
@@ -58,7 +58,7 @@ public class BackpackChangeMessageProcessor implements MessageProcessor<Backpack
                         .map(changeEntry -> {
                             try {
                                 final OdysseyMaterial odysseyMaterial = OdysseyMaterial.subtypeForName(changeEntry.getName());
-                                return new Pair<OdysseyMaterial,Long>(odysseyMaterial,changeEntry.getCount());
+                                return new Pair<OdysseyMaterial,Long>(odysseyMaterial,changeEntry.getCount().longValue());
                             } catch (final IllegalArgumentException e) {
                                 return null;
                             }
